+++ src/protocol/types/mod.rs
@@ -5,7 +5,9 @@
 pub mod contact;
 pub mod image;
 pub mod time;
+pub mod url;
 
 pub use self::contact::Contact;
 pub use self::image::ImageDimensions;
 pub use self::time::Timestamp;
+pub use self::url::RelayUrl;